refactor(3032): split model.rs's unit tests into model_tests.rs (campsite rule, epic #1116)
model.rs was already 883 lines (over the 800-line source target) before this
issue's bound-terminator change; extracting the test module leaves it at 455.
